[MODULARIZING-SCHEMA-RESOLVERS] Splitting schemaas/resolverrs code into resources specific files
