+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "apollo-server-micro": "^2.7.0-alpha.3",
     "graphql": "^14.4.1",
+    "lodash.merge": "^4.6.1",
     "micro": "^9.3.4",
     "microrouter": "^3.1.3"
   },
