+++ resources/owner/owner.resolvers.js
@@ -1,0 +1,11 @@
+const ownerMocked = require("../../mocks/owner");
+
+const resolvers = {
+  Query: {
+    owner() {
+      return ownerMocked;
+    }
+  }
+};
+
+module.exports = resolvers;
