+++ resources/resolvers.js
@@ -1,0 +1,3 @@
+const ownerResolvers = require("./owner/owner.resolvers");
+
+module.exports = ownerResolvers;
