+++ utils/dateFormatter.js
@@ -1,0 +1,18 @@
+const months = [
+  "JAN",
+  "FEB",
+  "MAR",
+  "APR",
+  "MAY",
+  "JUN",
+  "JUL",
+  "AUG",
+  "SEP",
+  "OCT",
+  "NOV",
+  "DEC"
+];
+
+module.exports = {
+  fromDateFormatter: date => `${months[date.getMonth()]} ${date.getFullYear()}`
+};
